requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
mlflow
dvc
python-dotenv
